[None][fix] Inline the FlashInfer decode re-plan gate; relocate its tests
Address review: inline the _records_fa2_plan helper into the decode-plan site
(if self.is_cuda_graph and decode_wrapper._backend != 'trtllm-gen') and simplify
the comment. The != 'trtllm-gen' test is intentional: a fa2-requested decode
wrapper is built with backend='auto' on non-SM90, which flashinfer resolves to a
concrete backend (possibly fa3), so == 'fa2' would miss it and skip recording the
plan under CUDA graphs. Move the plan-refresh behavior tests into
test_flashinfer_attention.py and drop the standalone file.
